% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Adapt.R
\name{adapt}
\alias{adapt}
\alias{is.Adapt}
\title{Define a Adaptive Procedure}
\usage{
adapt(
  delta = NA_real_,
  warmup = 0L,
  learn = "none",
  warmup_cov = -1L,
  batch = 1L,
  t0 = 10,
  kappa = 0.75,
  s_floor = 1e-04,
  s_ceil = 2.5,
  rm_gamma = 0.7,
  rm_c = 1,
  ridge = 1e-06,
  mode = "fast",
  mu = NA_real_,
  gamma = 0.2
)

is.Adapt(x)
}
\arguments{
\item{delta}{Numeric scalar in (0, 1). Target acceptance rate.}

\item{warmup}{Integer scalar. Number of warmup iterations}

\item{learn}{Character. Components of proposal to learn. Must be one of
's' (step size only), 'sD' (step size and shape), 'sDL' (step size, 
shape, and correlation), or 'none' (no adaptive procedure)}

\item{warmup_cov}{Integer scalar. Adaptive procedure is approximate D/L;
freeze D/L at warmup_cov; adapt s (only used for learn = 'sD' or learn = 'sDL')}

\item{batch}{Integer scalar in (0,1). Smooth over batch iterations}

\item{t0}{Positive numeric scalar. Stability; larger values essentially 
delay adaptation.}

\item{kappa}{Numeric scalar in (0,1). Smoothing; larger values are less 
smoothing.}

\item{s_floor}{Positive numeric scalar. Minimum allowed s.}

\item{s_ceil}{Positive numeric scalar. Maximum allowed s.}

\item{rm_gamma}{Positive numeric scalar > 0.5. Robbins-Monro lambda = c/ n^{gamma}}

\item{rm_c}{Positive numeric scalar. Robbins-Monro lambda = c/ n^{gamma}}

\item{ridge}{Positive numeric scalar. Ridge penalty in refresh.}

\item{mu}{Numeric scalar. The anchor point for the "fast" mode -- by
default set as log(10*s0)}

\item{gamma}{Numeric scalar. The gain term controlling aggressiveness
of the "fast" mode.}
}
\value{
An R6 \code{Adapt} object (class \code{"Adapt"}) with a stable
  \code{$toList()} representation suitable for downstream C++/Rcpp code.
}
\description{
Convenience constructor for a \code{Adapt} (R6) describing adaptation settings.
}
\section{Functions}{
\itemize{
\item \code{is.Adapt()}: Inheritance test

}}
\examples{
 adaptive = adapt(delta = 0.23, warmup = 2000L, batch = 5.0, t0 = 50, kappa = 0.75,
                 s_floor = 1e-3, s_ceil = 3.0, learn = 's')

}
