% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLatent.R
\name{addLatent}
\alias{addLatent}
\title{Add Latent Variable Draws from Chunk Files}
\usage{
addLatent(
  object,
  path,
  iterations = NULL,
  participants = NULL,
  pattern = NULL,
  as.fbm = FALSE,
  fbm.path = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{object}{A baclava object.}

\item{path}{A character string. Path to the directory containing latent
chunk files.}

\item{iterations}{An integer vector. Iteration indices to extract.
Note that iterations will be sorted before retrieval; returned object
may not be in the order provided here.}

\item{participants}{An optional logical or integer vector. If provided,
subsets participants (columns) from the latent matrices. Logical vectors
are applied directly; integer vectors are used as column indices.
Default \code{NULL} retains all participants.}

\item{pattern}{A character string. Regex pattern identifying chunk files.
Default assumes all *rds files in the directory are latent data files.}

\item{as.fbm}{A logical. If true, latent variables are returned as objects
of class FBM.}

\item{fbm.path}{A character string. If \code{as.fbm = TRUE}, the path in
which the FBM objects should be stored.}

\item{verbose}{Logical. Print progress messages. Default \code{TRUE}.}
}
\value{
The provided baclava object modified to include the latent data,
  and if iterations is specified, the posteriors are subset to align with
  the latent data.
}
\description{
Reads baclava latent variable chunk files from a single directory,
  extracting a specified subset of iterations and optionally a subset
  of participants. Returns the baclava object with posteriors align
  and latent element populated.
}
