% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aloocv.R
\name{aloocv}
\alias{aloocv}
\title{Approximate Leave-One-Out Cross-Validation}
\usage{
aloocv(
  object,
  data.clinical,
  data.assess,
  J.increment = 75L,
  J.max = 225L,
  ess.target = 50L,
  verbose = TRUE,
  lib = NULL
)
}
\arguments{
\item{object}{The value object returned by \link{fit_baclava}().}

\item{data.clinical}{A data.frame. The clinical data. The data must
 be structured as
 \itemize{
   \item \code{id}: A character, numeric, or integer object. The unique participant
     id to which the record pertains. Note these must include those provided in
     \code{data.assess}. Must be only 1 record for each participant.
   \item \code{age_entry}: A numeric object. The age at time of entry into the study.
     Note that this data is used to calculate a normalization; to expedite
     numerical integration, it is recommended that the ages be rounded.
     Optional input \code{round.age.entry} can be
     set to FALSE if this approximation is not desired; however, the
     computation time will significantly increase.
   \item \code{endpoint_type}: A character object. Must be one of \{"clinical",
     "censored", "preclinical"\}. Type "clinical" indicates that disease
     was diagnosed in the clinical compartment (i.e., symptomatic). Type
     "preclinical" indicates that disease was diagnosed in the preclinical
     compartment (i.e., during an assessment). Type
     "censored" indicates disease was not diagnosed prior to end of study.
   \item \code{age_endpoint}: A numeric object. The participant's age at the
     time the endpoint was evaluated.
}}

\item{data.assess}{A data.frame. Disease status assessments recorded
 during healthy or preclinical compartment, e.g., screenings for disease.
 The data must be structured as
 \itemize{
   \item \code{id}: A character, numeric, or integer object. The unique participant
   id to which the record pertains. Multiple records for each id are allowed.
   \item \code{age_assess}: A numeric object. The participant's age at time of
   assessment.
   \item \code{disease_detected}: An integer object. Must be binary 0/1, where
   1 indicates that disease was detected at the assessment; 0 otherwise.
}}

\item{J.increment}{An integer object. The number of replicates of each
participant to generate in each iteration of the importance sampling procedure
to attain desired effective sample size.}

\item{J.max}{An integer object. The maximum number of samples to be
drawn.}

\item{ess.target}{An integer object. The target effective sample size in the
importance sampling procedure.}

\item{verbose}{A logical object. If TRUE, progress information will be
printed. This input will be ignored if n.core > 1.}

\item{lib}{An optional character vector allowing for library path
to be provided to cluster.}
}
\value{
The provided baclava object augmented with the ALOOCV results. 
  Element aloocv is a list containing \code{summary}: the min, mean, and
  the 1% quantile of the ESS; the max, mean, and 99% quantile of J; the
  likelihood; and the individual-level and estimated predictive fit; and
  \code{result}: the likelihood, ESS, and J for each MCMC sample for each 
  participant.
}
\description{
Approximate leave-one-out cross-validation computed from the posterior
  draws of the Markov chain Monte Carlo sampler as implemented in
  \link{fit_baclava}().
}
\details{
Computes the predictive fit of a model.
  For each individual and each MCMC draw, the function approximates the
  marginal likelihood via importance sampling. It samples J.increment values
  of the individual's latent variables using the Metropolis-Hastings proposal
  distributions and computes the effective sample size (ESS) of the
  importance sampling procedure. If the target ESS is not met, J.increment
  additional samples are taken, and the ESS is re-evaluated.
  This is repeated until either the ESS is
  satisfied or J.max samples have been drawn.
}
\examples{

data(screen_data)

H <- weibull_dis(mu = MH(x = 100,
                         walk = reflective(1e-3, min = 0, log = TRUE),
                         prior = lnormp(log(100), 1)),
                 shape = fixed(2))

P <- weibull_dis(mu = MH(x = 2,
                         walk = reflective(1e-3, min = 0, log = TRUE),
                         prior = lnormp(log(2), 1)),
                 shape = fixed(2))

psi <- bern_dis(p = MH(x = .1,
                       walk = reflective(1e-3, min = 0.0, max = 1.0),
                       prior = betap(1, 1)))
                
beta <- bern_dis(p = Gibbs(x = .8, prior = betap(1, 1)))
                
# This is for illustration only -- the number of iterations should be
# significantly larger and the epsilon values should be tuned.
example <- fit_baclava(data.assess = data.screen,
                       data.clinical = data.clinical,
                       t0 = 30.0,
                       H = H, P = P, beta = beta, psi = psi)

res <- aloocv(example, data.clinical, data.screen)

}
