% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Model.R
\name{const_m}
\alias{const_m}
\alias{print.ConstantModel}
\title{A Constant Model}
\usage{
const_m(...)

\method{print}{ConstantModel}(x, ...)
}
\arguments{
\item{...}{Ignored. Used to require named inputs.}

\item{value}{An 'UpdateType' object with type = "MH" or "Fixed".}
}
\value{
An R6 object inheriting from class Model.
}
\description{
The parameter is modeled as \eqn{= value}{= value}.
}
\section{Functions}{
\itemize{
\item \code{print(ConstantModel)}: Print.

}}
\examples{

const_m(param(1.0, "r0", minv = 0, maxv = Inf, 
              prop_scale = "id", prior = lnormp(log(2), 0.2)))

}
