% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Distribution.R
\name{Distribution}
\alias{Distribution}
\alias{weibull_d}
\alias{gamma_d}
\alias{bern_d}
\alias{is.Distribution}
\alias{print.Distribution}
\title{Specification of Model Components}
\usage{
weibull_d(..., shape, rate, mu, scale, median)

gamma_d(..., shape, rate, mu, scale)

bern_d(..., p)

is.Distribution(x)

\method{print}{Distribution}(x, ...)
}
\arguments{
\item{...}{Ignored. Included to require named inputs.}

\item{shape}{A 'Param' object specifying the shape parameters. Use only
if Distribution is Weibull or Gamma.}

\item{rate}{If fitting the rate parameter of the distribution, a 'Param' or 
'Model' object. If fitting the mean or scale, do not specify. Use only
if Distribution is Weibull or Gamma.}

\item{mu}{If fitting the mean parameter of the distribution, a 'Param' or 
'Model' object. If fitting the rate or scale, do not specify. Use only
if Distribution is Weibull or Gamma.}

\item{scale}{If fitting the scale parameter of the distribution, a 'Param' or 
'Model' object. If fitting the rate or mu, do not specify. Use only
if Distribution is Weibull or Gamma.}

\item{p}{If fitting the probability parameter of the distribution, a 'Param' or 
'Model' object. Use only if Distribution is Bernoulli.}
}
\value{
An R6 object of class "Distribution".
}
\description{
The healthy and preclinical compartments of the model can be specified as
  Weibull or Gamma distributions. 
The screening sensitivity and indolent fraction are modeled as
  Bernoulli distributions.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(Distribution)}: Print

}}
\section{Functions}{
\itemize{
\item \code{weibull_d()}: Weibull Distribution

\item \code{gamma_d()}: Gamma Distribution

\item \code{bern_d()}: Bernoulli Distribution

\item \code{is.Distribution()}: Inheritance test

}}
