% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Model.R
\name{exp_m}
\alias{exp_m}
\alias{print.ExpModel}
\title{An Exponential Model}
\usage{
exp_m(..., r0, delta, m)

\method{print}{ExpModel}(x, ...)
}
\arguments{
\item{...}{Ignored. Used to require named inputs.}

\item{r0}{An 'UpdateType' object with type = "MH" or "Fixed".}

\item{delta}{An 'UpdateType' object with type = "MH" or "Fixed".}

\item{m}{An 'UpdateType' object with type = "MH" or "Fixed".}
}
\value{
An R6 object inheriting from class Model.
}
\description{
The parameter is modeled as 
\eqn{= r_0 \exp(\delta \times (\tau_{HP} - m))}{= r0 \exp(delta * (tau_HP - m))},
where \eqn{\tau_{HP}}{tau_HP} is the age at which a participant transitions
into the pre-clinical compartment.
}
\section{Functions}{
\itemize{
\item \code{print(ExpModel)}: Print.

}}
\examples{

# m is fixed
exp_m(r0 = param(1.0, "r0", minv = 0, maxv = Inf, 
                 prop_scale = "id", prior = lnormp(log(2), 0.2)),
      delta = param(1.0, "delta", prior = normp(2, 1)),
      m = param(52, "m"))

}
