% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_baclava.R, R/s3_methods.R
\name{fit_baclava}
\alias{fit_baclava}
\alias{summary.baclava}
\alias{print.baclava}
\title{Bayesian Analysis of Cancer Latency with Auxiliary Variable Augmentation}
\usage{
fit_baclava(
  data.assess,
  data.clinical,
  H,
  P,
  psi,
  beta,
  blocks,
  ...,
  baclava.object = NULL,
  M = 100L,
  thin = 1L,
  burnin = 0L,
  t0 = 0,
  indolent = TRUE,
  round.age.entry = TRUE,
  verbose = TRUE,
  save.latent = FALSE,
  debug = 0,
  m_point = Inf,
  prop_choice = 1,
  use_orig_prop = FALSE,
  I_marginalize = NULL,
  latent_dir = "./",
  latent_dump = 500L,
  file_prefix = "Analysis"
)

\method{summary}{baclava}(object, ..., burnin = 0L, thin = 0L)

\method{print}{baclava}(x, ...)
}
\arguments{
\item{data.assess}{A data.frame. Disease status assessments recorded
 during healthy or preclinical compartment, e.g., screenings for disease.
 The data must be structured as
 \itemize{
   \item \code{id}: A character, numeric, or integer object. The unique participant
   id to which the record pertains. Multiple records for each id are allowed.
   \item \code{age_assess}: A numeric object. The participant's age at time of
   assessment.
   \item \code{disease_detected}: An integer object. Must be binary 0/1, where
   1 indicates that disease was detected at the assessment; 0 otherwise.
}}

\item{data.clinical}{A data.frame. The clinical data. The data must
 be structured as
 \itemize{
   \item \code{id}: A character, numeric, or integer object. The unique participant
     id to which the record pertains. Note these must include those provided in
     \code{data.assess}. Must be only 1 record for each participant.
   \item \code{age_entry}: A numeric object. The age at time of entry into the study.
     Note that this data is used to calculate a normalization; to expedite
     numerical integration, it is recommended that the ages be rounded.
     Optional input \code{round.age.entry} can be
     set to FALSE if this approximation is not desired; however, the
     computation time will significantly increase.
   \item \code{endpoint_type}: A character object. Must be one of \{"clinical",
     "censored", "preclinical"\}. Type "clinical" indicates that disease
     was diagnosed in the clinical compartment (i.e., symptomatic). Type
     "preclinical" indicates that disease was diagnosed in the preclinical
     compartment (i.e., during an assessment). Type
     "censored" indicates disease was not diagnosed prior to end of study.
   \item \code{age_endpoint}: A numeric object. The participant's age at the
     time the endpoint was evaluated.
}}

\item{H}{An object of R6 class \code{Compartment} specifying the 
distribution(s) and parameter update strategy for the transition time 
through the Healthy compartment. The allowed distributions are currently
limited to weibull_d() and gamma_d().}

\item{P}{An object of R6 class \code{Compartment} specifying the 
distribution(s) and parameter update strategy for the transition time 
through the Preclinical compartment. The allowed distributions are currently
limited to weibull_d() and gamma_d().}

\item{psi}{An object of R6 class \code{Compartment} specifying the 
distribution(s) and parameter update strategy for the probability of 
indolence. The allowed distributions are currently
limited to bern_d().}

\item{beta}{An object of R6 class \code{Compartment} specifying the 
distribution(s) and parameter update strategy for the probability of 
screen detection. The allowed distributions are currently
limited to bern_d() and only Gibbs updates are allowed.}

\item{...}{Ignored.}

\item{baclava.object}{NULL or a 'baclava' object. To continue a calculation,
provide the object returned by a previous call.}

\item{M}{A positive integer object. The number of Monte Carlo samples. This
is the total, i.e., M = adaptive$warmup + n_MCMC.}

\item{thin}{A positive integer object. Keep each thin-th step of the sampler
after the warmup period, if any, is complete.}

\item{t0}{A non-negative scalar numeric object. The risk onset age. Must be
less than the earliest assessment age, entry age, and endpoint age.
If \code{baclava.object} is a 'baclava' object, this input is ignored.}

\item{indolent}{A logical object. If \code{FALSE}, disease under analysis
does not have an indolent state, i.e., it is always progressive.
This input is provided for convenience; if FALSE, psi is reset as
\code{compartment(distributions = list(bern_d(p = param(tag = "psi.p", init = 0.0))), strategies = NULL)}, 
which sets the value of psi to be 0 with no updates.
If \code{baclava.object} is a 'baclava' object, this input is ignored.}

\item{round.age.entry}{A logical object. If TRUE, the age at time of entry
will be rounded to the nearest integer prior to performing the MCMC.
This data is used to estimate the probability of experiencing clinical
disease prior to entering the study, which is estimated using a
time consuming numerical integration procedure. It is expected that
rounding the ages at time of entry introduces minimal bias. If FALSE,
and ages cannot be grouped, these integrals significantly increase
computation time.
If \code{baclava.object} is a 'baclava' object, this input is ignored.}

\item{verbose}{A logical object. If \code{TRUE}, a progress bar will be shown
during the MCMC.}

\item{save.latent}{A logical object. If \code{TRUE}, latent variable tau
and indolence will be returned. These can be very large matrices. To
estimate the cohort overdiagnosis probability using \link{cohortODX}(),
this must be set to TRUE.}

\item{object}{An object of class \code{baclava}.}

\item{x}{An object of class \code{baclava}.}
}
\value{
An object of S3 class \code{baclava}, which extends a list object.
  \itemize{
    \item theta: A list of the posterior distribution parameters at the 
      thinned samples.
    \item latent: A list of latent values and acceptance information
      \itemize{
        \item tau: If \code{save.latent = TRUE}, a matrix.
          The age at time of transition from
          healthy to preclinical compartment for each participant at the
          thinned samples. If \code{save.latent = FALSE}, their values at
          the last iteration.
.        \item indolent: If \code{save.latent = TRUE && I_marginalize != "ALL"}, a matrix.
          The indolent status for each participant at the
          thinned samples. If \code{save.latent = FALSE && I_marginalize != "ALL"}, 
          their values at
          the last iteration. Will be \code{NA} if disease is always progressive.
        \item tau_accept: If \code{save.latent = TRUE}, a matrix.
          TRUE/FALSE indicating if tau was accepted. Note that NA is used
          to indicate an Inf->Inf transition. 
          If \code{save.latent = FALSE}, their values at
          the last iteration.  
        \item accept_rate: Individual level acceptance rates for tau
       }
    \item accept: A list of the accept indicator of each MH step at the thinned samples.
    \item acceptance_prob A list of the acceptance probabilities in the post-burnin
      period of the sampler for each MH step.
    \item epsilon: A list. The random walk settings used in the post-warmup period.
    \item adaptive: A list. The random walk values at each iteration of the
      adaptive period.
    \item setup: A list of inputs provided to the call.
      \itemize{
        \item t0: The input age of risk onset.
        \item indolent: TRUE if disease is not progressive.
        \item round.age.entry: TRUE if age at entry was rounded to the nearest
          whole number.
        \item M The total number of iterations.
        \item thin: The number of samples dropped between kept MCMC iterations.
        \item burnin: The number of iterations dropped.
     }
    \item model: The provided model specification.
    \item call: The matched call.
  }
}
\description{
Markov chain Monte Carlo sampler to fit a three-state mixture compartmental
 model of cancer natural history to individual-level screening and cancer
 diagnosis histories in a Bayesian framework.
}
\section{Functions}{
\itemize{
\item \code{summary(baclava)}: Summary statistics of posterior distribution parameters

\item \code{print(baclava)}: Print summary statistics of posterior distribution parameters

}}
\examples{

data(screen_data)

# H compartment model
# fit parameter is the mean of the distribution
mu_H <- param(100.0, tag = "mu_H", minv = 0.0, maxv = Inf,
              prior = lnorm_p(log(100), 1))
# shape parameter
sh_H <- param(2.0, tag = "sh_H"), minv =  0.0, maxv = Inf
               prior = lnorm_p(log(2), 0.1))
# Weibull distribution
H <- compartment(weibull_d(mean = mu_H, shape = sh_H))

# P compartment model
# fit parameter is the mean of the distribution
mu_P <- param(2, tag = "mu_P", minv = 0.0, maxv = Inf,
              prior = lnorm_p(log(2), 1))
# shape parameter
sh_P <- param(2.0, tag = "sh_P"), minv =  0.0, maxv = Inf
               prior = lnorm_p(log(2), 0.1))
# Weibull distribution
P <- compartment(weibull_d(mean = mu_P, shape = sh_P))

# psi 
# probability parameter
p_psi <- param(0.2, tag = "p_psi", minv = 0.0, maxv = 1.0, 
               prior = beta_p(1.0, 1.0))

psi <- compartment(bern_d(p = p_psi))

# beta 
p_b <- param(0.8, tag = "p_b", minv = 0.0, maxv = 1.0, prior = beta_p(1, 1))
beta <- compartment(bern_d(p = p_b))

#Sampler strategies
strategy_H <- MH(params = c("sc_H"), s0 = 0.014, 
                 adaptive = adapt(learn = "none"))

strategy_P <- MH(params = c("r_P"), s0 = 0.06, 
                 adaptive = adapt(learn = "none"))

strategy_psi <- MH(params = c("p_psi"), s0 = 0.1,
                   adaptive = adapt(learn = "none"))
blocks <- list(strategy_H, strategy_P, Gibbs("p_b"), strategy_psi)

# This is for illustration only -- the number of iterations should be
# significantly larger and the epsilon values should be tuned.
example <- fit_baclava(data.assess = data.screen,
                       data.clinical = data.clinical,
                       t0 = 30.0,
                       H = H, P = P, beta = beta, psi = psi,
                       blocks = blocks)

summary(example)
print(example)

# To continue this calculation
example_continued <- fit_baclava(data.assess = data.screen,
                                 data.clinical = data.clinical,
                                 baclava.object = example)
}
