% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Strategy.R
\name{Gibbs}
\alias{Gibbs}
\alias{is.Gibbs}
\title{Define a single Gibbs update}
\usage{
Gibbs(params)

is.Gibbs(x)
}
\arguments{
\item{params}{Character vector of parameter names included in the Gibbs Step.}
}
\value{
An R6 \code{Gibbs} object (class \code{"Gibbs"}) with a stable
  \code{$toList()} representation suitable for downstream C++/Rcpp code.
}
\description{
Convenience constructor for a \code{Gibbs} (R6) describing one MCMC Gibbs update
step. Use this to specify which parameters are updated together.
}
\section{Functions}{
\itemize{
\item \code{is.Gibbs()}: Inheritance test

}}
\examples{
# Gibbs over two sensitivity parameters
b1 <- Gibbs(params = c("beta0", "beta1"))

}
