% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{plot.baclava}
\alias{plot.baclava}
\title{Plot Posterior Distribution Parameters}
\usage{
\method{plot}{baclava}(
  x,
  y,
  ...,
  type = c("density", "trace", "acf"),
  burnin = 0L,
  thin = 0L,
  trace_var = NULL,
  scale = c("natural", "proposal")
)
}
\arguments{
\item{x}{An object of class \code{baclava}.}

\item{y}{Ignored}

\item{...}{Ignored}

\item{type}{A character object. One of \{"density", "trace", "acf"\}. The
type of plot to generate}

\item{burnin}{An integer object. Optional. The number of burn-in samples.
Used only for \code{type = "trace"}. One trace plot is generated for
the burnin iterations; a second for the post-burnin iterations. Note, this
refers to the number of kept (thinned) samples.}

\item{thin}{An integer object. If non-zero posterior will be thinned to
n, 2n, 3n, ...}

\item{trace_var}{A character object. The parameter for which trace plots
are to be generated.}
}
\value{
A gg object
}
\description{
Convenience function to facilitate exploration of posterior distributions through
  trace plots, autocorrelations, and densities.
}
\examples{
data(screen_data)

# H compartment model

# fitted parameter is the mean of the distribution
mu_H <- param(100.0, tag = "mu_H", minv = 0.0, maxv = Inf,
              prior = lnorm_p(log(100), 1))
              
# fixed shape parameter
sh_H <- param(2.0, tag = "sh_H")

# Weibull distribution
H <- compartment(weibull_d(mean = mu_H, shape = sh_H))

# P compartment model

# fitted parameter is the mean of the distribution
mu_P <- param(2, tag = "mu_P", minv = 0.0, maxv = Inf,
              prior = lnorm_p(log(2), 1))
              
# fixed shape parameter
sh_P <- param(2.0, tag = "sh_P")

# Weibull distribution
P <- compartment(weibull_d(mean = mu_P, shape = sh_P))

# psi 
# probability parameter
p_psi <- param(0.2, tag = "p_psi", minv = 0.0, maxv = 1.0, 
               prior = beta_p(1.0, 1.0))

psi <- compartment(bern_d(p = p_psi))

# beta 
p_b <- param(0.8, tag = "p_b", minv = 0.0, maxv = 1.0, prior = beta_p(1, 1))
beta <- compartment(bern_d(p = p_b))

#Sampler strategies
strategy_H <- MH(params = c("mu_H"), s0 = 0.014, 
                 adaptive = adapt(learn = "none"))

strategy_P <- MH(params = c("mu_P"), s0 = 0.06, 
                 adaptive = adapt(learn = "none"))

strategy_psi <- MH(params = c("p_psi"), s0 = 0.1,
                   adaptive = adapt(learn = "none"))
blocks <- list(strategy_H, strategy_P, Gibbs("p_b"), strategy_psi)

# This is for illustration only -- the number of iterations should be
# significantly larger and the epsilon values should be tuned.
example <- fit_baclava(data.assess = data.screen,
                       data.clinical = data.clinical,
                       t0 = 30.0,
                       H = H, P = P, beta = beta, psi = psi,
                       blocks = blocks)
                       
plot(example)
plot(example, type = "trace", trace_var = "p_psi", burnin = 0L) 
plot(example, type = "trace", trace_var = "mu_H", burnin = 0L) 
plot(example, type = "trace", trace_var = "mu_P", burnin = 0L) 
plot(example, type = "trace", trace_var = "p_b", burnin = 0L) 
plot(example, type = "acf")

}
