% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictODX.R, R/predictODX_old.R
\name{predictODX}
\alias{predictODX}
\alias{plot.baclava.ODX.pred}
\alias{print.baclava.ODX.pred}
\alias{predictODX_old}
\alias{plot.baclava.ODX.pred.old}
\title{Estimate the Overall and Per Screen Overdiagnosis Rates}
\usage{
predictODX(
  object,
  screening.schedule,
  other.cause.rates,
  model = list(H = "all", P = "all", psi = "all", beta = "all"),
  burnin = 0L,
  thin = 1L,
  verbose = TRUE,
  n.draws = NULL,
  n.core = max(1L, parallel::detectCores() - 1L),
  ...
)

\method{plot}{baclava.ODX.pred}(x, y, ...)

\method{print}{baclava.ODX.pred}(x)

predictODX_old(
  object,
  screening.schedule,
  other.cause.rates,
  n.draws = 1000L,
  burnin = 1000L,
  verbose = TRUE
)

\method{plot}{baclava.ODX.pred.old}(x, y, ...)
}
\arguments{
\item{object}{An object of S3 class 'baclava'. The value object returned
by \code{fit_baclava()}.}

\item{screening.schedule}{A numeric vector object. A vector of ages at
which screenings occur.}

\item{other.cause.rates}{A data.frame object. Must contain columns "Rate"
and "Age".}

\item{model}{A named list. If the original model assumed a subset structure,
mixed preclinical compartment, or multiple beta models,
this list specifies the model component to use to predict ODX.
If any model components are subset specific,
only 1 subset can be specified. For mixture models or multiple beta models 
apply to the subset, provide names of all needed models.  
Names of the provided list elements must be 'H', 'P', 'psi' and 'beta'.
The options available can be seen by printing names(object$model$X$distributions) where
X is one of P, H, psi, or beta.}

\item{burnin}{An integer object. Optional. The number of burn-in samples.
Used only for \code{type = "trace"}. One trace plot is generated for
the burnin iterations; a second for the post-burnin iterations. Note, this
refers to the kept (thinned) samples.}

\item{thin}{An integer object. Optional. If > 1, posterior samples 
(after removing burnin samples) are thinned to (thin, 2*thin, 3*thin, ...)}

\item{verbose}{A logical object. If TRUE, progress bars will be displayed.}

\item{n.core}{A scalar integer. The number of cores for parallelization (=1 is sequential)}

\item{...}{Ignored.}

\item{x}{A an object of S3 class 'baclava.PDX.pred' as returned by \code{predictODX()}.}

\item{y}{Ignored.}

\item{groups.rateP}{An integer scalar object. If model included groups with
different sojourn parameters, the group for which overdiagnosis is to
be estimated. Must be one of \code{object$setup$groups.rateP}}

\item{screen.type}{An integer scalar object. If model included screen-type,
specific sensitivity parameters, the screen-type for which
overdiagnosis is to be estimated. Must be one of \code{object$setup$groups.beta}}
}
\value{
A list object. For each screen in \code{screening.schedule},
  a matrix providing the mean total overdiagnosis and the mean overdiagnosis
  due to indolent/progressive tumors, as well as their 95\% prediction intervals.
  Similarly, element \code{overall} provides these estimates for the full
  screening schedule.

A list object. For each screen in \code{screening.schedule},
  a matrix providing the mean total overdiagnosis and the mean overdiagnosis
  due to indolent/progressive tumors, as well as their 95% prediction intervals.
  Similarly, element \code{overall} provides these estimates for the full
  screening schedule.
}
\description{
Using the posterior parameter distributions, calculates the infinite
  population estimates of the probability of overdiagnosis at 
  each screening episode due to indolence and/or death by other causes.

Using the posterior parameter distributions, calculates the infinite
  population estimates of the probability of overdiagnosis at 
  each screening episode due to indolence and/or death by other causes.
}
\details{
Provided birth cohort life table is an all cause tables obtained from the 
  CDC Life Tables. Vital Statistics of the United States, 1974 Life Tables, 
  Vol. II, Section 5. 1976. Estimated "other cause" mortality will thus be 
  overestimated when using these tables. It is recommended that user provide 
  data that has been corrected to exclude death due to the disease under analysis.

Provided birth cohort life table is an all cause tables obtained from the 
  CDC Life Tables. Vital Statistics of the United States, 1974 Life Tables, 
  Vol. II, Section 5. 1976. Estimated "other cause" mortality will thus be 
  overestimated when using these tables. It is recommended that user provide 
  data that has been corrected to exclude death due to the disease under analysis.
}
\section{Functions}{
\itemize{
\item \code{plot(baclava.ODX.pred)}: Generate column plot of predicted overdiagnosis for each screen.

\item \code{print(baclava.ODX.pred)}: print method

\item \code{plot(baclava.ODX.pred.old)}: Generate column plot of predicted overdiagnosis for each screen.

}}
\examples{

data(screen_data)

# H compartment model

mu_H <- param(init = runif(1L, 50.0, 500.0), tag = "mu_H", 
              minv = 10.0, prop_scale = "log", 
              prior = lnorm_p(log(150.00), 0.4))
sh_H <- param(init =  runif(1L, 0.5, 5.0), tag = "sh_H", 
              minv =  0.01, prop_scale = "log", 
              prior = lnorm_p(log( 1.75), 0.4))
dist_H <- weibull_d(mu = mu_H, shape = sh_H)
strategy_H <- block(params = c("mu_H", "sh_H"), adaptive = list(warmup = 0))

H <- compartment(dist_H, strategy_H)

# beta model

p_b <- param(init = runif(1L, 0.0, 1.0), tag = "p_b", 
             minv = 0.0, maxv = 1.0, 
             prior = beta_p(1.0, 1.0))
dist_b <- bern_d(p = p_b)

beta <- compartment(dist_b, strategies = "Gibbs")

# P compartment model

mu_P <- param(init = runif(1L, 1.0, 10.0), tag = "mu_P", 
              minv = 1e-6, prop_scale = "log", 
              prior = lnorm_p(log(2), 0.4))
sh_P <- param(init = runif(1L, 0.5, 5.0), tag = "sh_P", 
              minv =  0.01, prop_scale = "log", 
              prior = lnorm_p(log(1.73), 0.28))
dist_P <- gamma_d(mu = mu_P, shape = sh_P)
strategy_P <- block(params = c("mu_P", "sh_P"), adaptive = list(warmup = 0))

P <- compartment(dist_P, list(strategy_P))

# psi model

p_psi <- param(init = runif(1L, 0.0, 1.0), tag = "p_psi", 
               minv = 0.0, maxv = 1.0, 
               prior = beta_p(1.0, 1.0))
dist_psi <- bern_d(p = p_psi)
strategy_psi <- block(params = c("p_psi"), adaptive = list(warmup = 0))

psi <- compartment(dist_psi, blockStrategies(list(strategy_psi)))

# This is for illustration only -- the number of iterations should be
# significantly larger and the epsilon values should be tuned.
example <- fit_baclava(data.assess = sim_screen,
                       data.clinical = sim_clinical,
                       t0 = t0,
                       H = H, P = P, psi = psi, beta = beta)
                
# if rates are not available, an all cause dataset is provided in the package
# NOTE: these predictions will be over-estimated
           
data(all_cause_rates)
all_cause_rates <- all_cause_rates[, c("Age", "both")]
colnames(all_cause_rates) <- c("Age", "Rate")

# using single screen for example speed
predicted_odx <- predictODX(object = example, 
                            other.cause.rates = all_cause_rates,
                            screening.schedule = 40, 
                            burnin = 10)

plot(predicted_odx)


data(screen_data)

theta_0 <- list("rate_H" = 7e-4, "shape_H" = 2.0,
                "rate_P" = 0.5  , "shape_P" = 1.0,
                "beta" = 0.9, psi = 0.4)
prior <- list("rate_H" = 0.01, "shape_H" = 1,
              "rate_P" = 0.01, "shape_P" = 1,
              "a_psi" = 1/2 , "b_psi" = 1/2,
              "a_beta" = 38.5, "b_beta" = 5.8)

# This is for illustration only -- the number of MCMC samples should be
# significantly larger and the epsilon values should be tuned.
example <- fit_baclava(data.assess = data.screen,
                       data.clinical = data.clinical,
                       t0 = 30.0,
                       theta_0 = theta_0,
                       prior = prior)
                       
# if rates are not available, an all cause dataset is provided in the package
# NOTE: these predictions will be over-estimated
           
data(all_cause_rates)
all_cause_rates <- all_cause_rates[, c("Age", "both")]
colnames(all_cause_rates) <- c("Age", "Rate")

# using single screen for example speed
predicted_odx <- predictODX(object = example, 
                            other.cause.rates = all_cause_rates,
                            screening.schedule = 40, 
                            burnin = 10)

plot(predicted_odx)

}
