% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Prior.R
\name{priors}
\alias{priors}
\alias{unif_p}
\alias{norm_p}
\alias{halfnorm_p}
\alias{lnorm_p}
\alias{gamma_p}
\alias{invgamma_p}
\alias{beta_p}
\alias{nbeta_p}
\alias{chisq_p}
\alias{nchisq_p}
\alias{t_p}
\alias{nt_p}
\alias{f_p}
\alias{nf_p}
\alias{cauchy_p}
\alias{exp_p}
\alias{logis_p}
\alias{weibull_p}
\alias{binom_p}
\alias{prior_p}
\alias{print.Prior}
\title{Prior Specifications}
\usage{
unif_p(min = 0, max = 1)

norm_p(mean = 0, sd = 1)

halfnorm_p(sd = 1)

lnorm_p(meanlog = 0, sdlog = 1)

gamma_p(shape, scale = 1)

invgamma_p(shape, scale = 1)

beta_p(shape1, shape2)

nbeta_p(shape1, shape2, ncp)

chisq_p(df)

nchisq_p(df, ncp)

t_p(df)

nt_p(df, ncp)

f_p(df1, df2)

nf_p(df1, df2, ncp)

cauchy_p(location = 0, scale = 1)

exp_p(rate = 1)

logis_p(location = 0, scale = 1)

weibull_p(shape, scale = 1)

binom_p(size, prob)

prior_p(dist, ...)

\method{print}{Prior}(x, ...)
}
\arguments{
\item{min, max}{The lower and upper limits of the distribution. Must be finite.}

\item{mean, sd}{The mean and standard deviation of the distribution.}

\item{sd}{The standard deviation of the distribution.}

\item{meanlog, sdlog}{The mean and standard deviation of the distribution on the log scale.}

\item{shape, scale}{The shape and scale parameters. Must be positive, scale strictly.}

\item{shape1, shape2}{The non-negative parameters.}

\item{ncp}{The non-centrality parameter.}

\item{df}{The degrees of freedom; must be non-negative.}

\item{df1, df2}{The degrees of freedom; must be non-negative.}

\item{location, scale}{The location and scale parameters.}

\item{rate}{The rate parameter.}
}
\value{
An S3 object of class `Prior`, which extends list. Element `dist`
  is the character name of the density function and `params` is a numeric
  vector of the distribution parameters.
}
\description{
Convenience functions for specifying the prior distributions and their
  parameter settings for each model parameter.
}
\section{Functions}{
\itemize{
\item \code{norm_p()}: Normal Distribution

\item \code{halfnorm_p()}: Half-Normal Distribution

\item \code{lnorm_p()}: Log-Normal Distribution

\item \code{gamma_p()}: Gamma Distribution

\item \code{invgamma_p()}: Inverse-Gamma Distribution

\item \code{beta_p()}: Beta Distribution

\item \code{nbeta_p()}: Noncentral Beta Distribution

\item \code{chisq_p()}: Chi-squared Distribution

\item \code{nchisq_p()}: Noncentral Chi-squared Distribution

\item \code{t_p()}: T Distribution

\item \code{nt_p()}: Noncentral T Distribution

\item \code{f_p()}: F Distribution

\item \code{nf_p()}: Noncentral F Distribution

\item \code{cauchy_p()}: Cauchy Distribution

\item \code{exp_p()}: Exponential Distribution

\item \code{logis_p()}: Logistic Distribution

\item \code{weibull_p()}: Weibull Distribution

\item \code{binom_p()}: Binomial Distribution

\item \code{prior_p()}: General Constructor

\item \code{print(Prior)}: Print

}}
\examples{

# prior is dunif(x, min = 0, max = 10)
unif_p(0, 10)

# prior is dnorm(x, mean = 0.2, sd = 1.5)
norm_p(0.2, 1.5)

# prior is abs(dnorm(x, mean = 0.0, sd = 1.5))
halfnorm_p(1.5)

# prior is dlnorm(x, meanlog = 1, sdlog = 1.5)
lnorm_p(1, 1.5)

# prior is dgamma(x, shape = 1, scale = 1.5)
gamma_p(1, 1.5)

# prior is dinvgamma(x, shape = 1, scale = 1.5)
invgamma_p(1, 1.5)

# prior is dbeta(x, shape1 = 1, shape2 = 2)
beta_p(1, 2)

# prior is dbeta(x, shape1 = 1, shape2 = 1.5, ncp = 0.4)
nbeta_p(1, 1.5, 0.4)

# prior is dchisq(x, df = 2)
chisq_p(2)

# prior is dchisq(x, df = 2, ncp = 0.4)
nchisq_p(2, 0.4)

# prior is dt(x, df = 2)
t_p(2)

# prior is dt(x, df = 2, ncp = 0.4)
nt_p(2, 0.4)

# prior is df(x, df1 = 2, df2 = 4)
f_p(2, 4)

# prior is df(x, df1 = 2, df2 = 4, ncp = 0.4)
nf_p(2, 4, 0.4)

# prior is dcauchy(location = 0.4, scale = 1.4)
cauchy_p(0.4, 1.4)

# prior is dexp(rate = 1.4)
exp_p(1.4)

# prior is dlogis(location = 0.4, scale = 1.4)
logis_p(0.4, 1.4)

# prior is dweibull(x, shape = 0.4, scale = 1.4)
weibull_p(0.4, 1.4)

# prior is dbinom(x, size = 1, prob = 0.5)
binom_p(1, 0.5)

# prior is dbinom(x, 1, 0.5)
prior_p("dbinom", 1, 0.5)

# prior is dweibull(x, shape = 0.4, scale = 1.4)
prior_p("dweibull", 0.4, 1.4)
}
