% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Model.R
\name{sigmoid_m}
\alias{sigmoid_m}
\alias{print.SigmoidModel}
\title{A Sigmoid Model}
\usage{
sigmoid_m(..., r0, delta, m, w)

\method{print}{SigmoidModel}(x, ...)
}
\arguments{
\item{...}{Ignored. Used to require named inputs.}

\item{r0}{An 'UpdateType' object with type = "MH" or "Fixed".}

\item{delta}{An 'UpdateType' object with type = "MH" or "Fixed".}

\item{m}{An 'UpdateType' object with type = "MH" or "Fixed".}

\item{w}{An 'UpdateType' object with type = "MH" or "Fixed".}
}
\value{
An R6 object inheriting from class Model.
}
\description{
The parameter is modeled as
\eqn{= r_0 + \delta / (1 + exp(-(\tau_{HP} - m) / w))}{= r0 + delta / (1 + exp(-(tau_HP - m) / w},
where \eqn{\tau_{HP}}{tau_HP} is the age at which a participant transitions
into the pre-clinical compartment.
}
\section{Functions}{
\itemize{
\item \code{print(SigmoidModel)}: Print.

}}
\examples{

sigmoid_m(r0 = param(1.0, "r0", minv = 0, maxv = Inf, 
                     prop_scale = "id", prior = lnormp(log(2), 0.2)),
          delta = param(1.0, "delta", prior = normp(2, 1)),
          m = param(52, "m"),
          w = param(1, "w", minv = 0, maxv = Inf, prior = lnormp(log(1), 0.2)))
          
}
