% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Model.R
\name{step_m}
\alias{step_m}
\alias{print.StepModel}
\title{A Step Model}
\usage{
step_m(..., r0, r1, m)

\method{print}{StepModel}(x, ...)
}
\arguments{
\item{...}{Ignored. Used to require named inputs.}

\item{r0}{An 'UpdateType' object with type = "MH" or "Fixed".}

\item{r1}{An 'UpdateType' object with type = "MH" or "Fixed".}

\item{m}{An 'UpdateType' object with type = "MH" or "Fixed".}
}
\value{
An R6 object inheriting from class Model.
}
\description{
The parameter is modeled as 
\eqn{= r_0 (\tau_{HP} < m) + r_1 (\tau_{HP} >= m)}{= r_0 (\tau_{HP} < m) + r_1 (\tau_{HP} >= m)}
where \eqn{\tau_{HP}}{tau_HP} is the age at which a participant transitions
into the pre-clinical compartment.
}
\section{Functions}{
\itemize{
\item \code{print(StepModel)}: Print.

}}
\examples{

# m is fixed
step_m(r0 = param(1.0, "r0", minv = 0, maxv = Inf, 
                  prop_scale = "id", prior = lnormp(log(2), 0.2)),
       r1 = param(1.0, "r0", minv = 0, maxv = Inf, 
                  prop_scale = "id", prior = lnormp(log(2), 0.2)),
       m = param(52, "m"))
          
}
